refactored to put functions in the right module; collapsed tests into one
